Moved location of EditPasswordModal.js and edited buttons for Profile.js
